requirements.txt
flask
python-dotenv
requests
redis
redis-server
